cleanup - lift shared wrapper components + remove duplicates
