+++ packages/renderer/src/renderer/screens/Import/Ledger/Import/InfoCardSteps.tsx
@@ -9,7 +9,7 @@ import {
   faCircleChevronRight,
   faCircleDot,
 } from '@fortawesome/free-solid-svg-icons';
-import { InfoCardStepsWrapper } from './Wrappers';
+import { InfoCardStepsWrapper } from '../../Wrappers';
 
 /**
  * TODO: Move component to library.
